% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleangeo.R
\docType{package}
\name{cleangeo}
\alias{cleangeo}
\alias{cleangeo-package}
\title{A Set of Utility Tools to Clean Geometries}
\description{
cleangeo provides you a set of utility tools to inspect spatial objects,
report errors by type and clean geometries.
}
\details{
\tabular{ll}{
   Package: \tab cleangeo\cr
   Type: \tab Package\cr
   Version: \tab 0.1-0\cr
   Date: \tab 2015-09-25\cr
   License: \tab GPL(>=2.0)\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

