% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clgeo_CleanByPolygonation.R
\name{clgeo_CleanByPolygonation.SpatialPolygons}
\alias{clgeo_CleanByPolygonation.SpatialPolygons}
\alias{clgeo_CleanByTriangulation}
\title{clgeo_CleanByPolygonation.SpatialPolygons}
\usage{
clgeo_CleanByPolygonation.SpatialPolygons(sp, verbose = FALSE)
}
\arguments{
\item{sp}{object extending the \code{\link[sp]{Spatial-class}} as defined in \pkg{sp}}

\item{verbose}{Indicates wether the clean logs have to be printed. Default value is FALSE.}
}
\value{
an object extending the \code{\link[sp]{Spatial-class}} as defined in \pkg{sp},
with cleaned geometries.
}
\description{
Function to clean a \code{\link[sp]{SpatialPolygons}} object by polygonation
}
\note{
The polygonation method is a tentative alternate method to triangulation to clean
geometries. In this method, triangulation is skipped and a re-polygonation algorithm is
applied.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

