% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clgeo_CleanByTriangulation.R
\name{clgeo_CleanByTriangulation.Polygon}
\alias{clgeo_CleanByTriangulation.Polygon}
\title{clgeo_CleanByTriangulation.Polygon}
\usage{
clgeo_CleanByTriangulation.Polygon(p)
}
\arguments{
\item{p}{object of class \code{\link[sp]{Polygon-class}} as defined in \pkg{sp}}
}
\value{
a list of objects of class \code{\link[sp]{Polygon-class}} as defined in \pkg{sp},
with cleaned geometries.
}
\description{
Function to clean a \code{\link[sp]{Polygon-class}} object by triangulation
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

