% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clgeo_CleanByTriangulation.R
\name{clgeo_CleanByTriangulation.Polygons}
\alias{clgeo_CleanByTriangulation}
\alias{clgeo_CleanByTriangulation.Polygons}
\title{clgeo_CleanByTriangulation.Polygons}
\usage{
clgeo_CleanByTriangulation.Polygons(p)
}
\arguments{
\item{polygons}{object of class \code{\link[sp]{Polygons-class}} as defined in \pkg{sp}}
}
\value{
an object of class \code{\link[sp]{Polygons-class}} as defined in \pkg{sp},
with cleaned geometries.
}
\description{
Function to clean a \code{\link[sp]{Polygons}} object by triangulation
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

