% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clgeo_CleanByTriangulation.R
\name{clgeo_CleanByTriangulation.SpatialPolygons}
\alias{clgeo_CleanByTriangulation}
\alias{clgeo_CleanByTriangulation.SpatialPolygons}
\title{clgeo_CleanByTriangulation.SpatialPolygons}
\usage{
clgeo_CleanByTriangulation.SpatialPolygons(sp)
}
\arguments{
\item{sp}{object extending the \code{\link[sp]{Spatial-class}} as defined in \pkg{sp}}
}
\value{
an object extending the \code{\link[sp]{Spatial-class}} as defined in \pkg{sp},
with cleaned geometries.
}
\description{
Function to clean a \code{\link[sp]{SpatialPolygons}} object by triangulation
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

