% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgeo_GeometryReport.R
\name{clgeo_GeometryReport}
\alias{clgeo_GeometryReport}
\alias{clgeo_Geometry}
\title{clgeo_GeometryReport}
\usage{
clgeo_GeometryReport(spgeom)
}
\arguments{
\item{spgeom}{object extending the \code{\link[sp]{Spatial-class}} as defined in \pkg{sp}}
}
\value{
an object of class \code{list} giving the following:
\itemize{
  \item \emph{type} eventual \pkg{rgeos} issue
  \item \emph{valid} geometry validity status (according to OGC specifications)
  \item \emph{issue_type} type of geometry issue
  \item \emph{msg} catched message when warning raised about geometry
}
}
\description{
Function to get a geometry validation report: The report informs on the following:

\itemize{
  \item \emph{type} eventual \pkg{rgeos} issue
  \item \emph{valid} geometry validity status (according to OGC specifications)
  \item \emph{issue_type} type of geometry issue
  \item \emph{msg} catched message when error raised about geometry
}
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{geometry}
\keyword{validity}
