% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleangeo.R
\docType{package}
\name{cleangeo}
\alias{cleangeo}
\alias{cleangeo-package}
\title{Clean Geometries from Spatial Objects}
\description{
cleangeo provides a set of utility tools to inspect spatial objects, facilitate handling and
reporting of topology errors and geometry validity issues.  Finally, it provides a geometry
cleaner that will fix all geometry problems, and  eliminate (at least reduce) the likelihood
of having issues when doing spatial data processing.
}
\details{
\tabular{ll}{
   Package: \tab cleangeo\cr
   Type: \tab Package\cr
   Version: \tab 0.2\cr
   Date: \tab 2016-07-06\cr
   License: \tab GPL(>=2.0)\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

