% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleangeo.R
\docType{package}
\name{cleangeo}
\alias{cleangeo}
\alias{cleangeo-package}
\title{Clean Geometries from Spatial Objects}
\description{
cleangeo provides a set of utility tools to inspect spatial objects, facilitate handling and 
reporting of topology errors and geometry validity issues.  Finally, it provides a geometry 
cleaner that will fix all geometry problems, and  eliminate (at least reduce) the likelihood 
of having issues when doing spatial data processing.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
