% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clgeo_Clean}
\alias{clgeo_Clean}
\title{clgeo_Clean}
\usage{
clgeo_Clean(sp, errors.only = NULL, print.log = TRUE)
}
\arguments{
\item{sp}{object extending the \code{\link[sp]{Spatial-class}}
as defined in \pkg{sp}}

\item{errors.only}{an object of class \code{vector} giving the types of errors
for which the output should bounded. Default value is NULL (\emph{i.e.} the output
will include features for which both errors and errors were raised.). At now, this
argument accepts the error type \code{"ORPHANED_HOLE"}.}

\item{print.log}{Indicates wether the clean logs have to be printed. Default
value is TRUE.}
}
\value{
an object extending the \code{\link[sp]{Spatial-class}}
as defined in \pkg{sp}, with cleaned geometries.
}
\description{
Function to clean a spatial data collection
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

