% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clgeo_CleanByPolygonation.R
\name{clgeo_CleanByPolygonation.Polygon}
\alias{clgeo_CleanByPolygonation.Polygon}
\title{clgeo_CleanByPolygonation.Polygon}
\usage{
clgeo_CleanByPolygonation.Polygon(p)
}
\arguments{
\item{p}{object of class \code{\link[sp]{Polygon-class}} as defined in \pkg{sp}}
}
\value{
a list of objects of class \code{\link[sp]{Polygon-class}} as defined in \pkg{sp},
with cleaned geometries.
}
\description{
Function to clean a \code{\link[sp]{Polygon-class}} object by polygonation.
}
\note{
The polygonation method is a tentative alternate method to triangulation to clean
geometries. In this method, triangulation is skipped and a re-polygonation algorithm is
applied.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
\keyword{clean}
\keyword{geometry}
\keyword{summary}
\keyword{validity}

